Fix raw markdown showing in Phase Milestone Exam and exercise questions
MasteryCheck and ExerciseWidget rendered question text, answers, and
instructions by naively splitting on newlines into <p> tags, so bold
text, inline code, and fenced code blocks appeared as literal markdown
syntax instead of rendering. Extract the shared ReactMarkdown plugin
config from MarkdownRenderer into MarkdownFragment and use it in both
components.

Also fixes a related bug where the mastery-check <details> hint/answer
block was read from only the first HTML fragment produced by the
markdown parser (which splits <details> at blank lines), leaving the
answer empty for every question with a multi-line hint list.
